Added tx.close() call to TimeOp.java
TimeOp.java was relying on GC of the RAMCloudTransaction object to cleanup the
C++ state via the RAMCloudTransaction.finalize() method called by the GC... but
this was leading to a segmentation fault after executing on the order of 10k tx
reads. While I don't know for sure, I suspect that two issues with this approach
may be causing the segfault:

1) The Java Finalizer thread may not call finalizer for a long time. This leads
to memory getting filled up. However, from the hs_err_pidXXX.txt files, I do not
see anything indicating that the JVM ran out of memory.

2) The Java Finalizer thread runs in parallel with the rest of the code, but the
RAMCloud Java libraries (and the RAMCloud C++ client library) are not thread
safe. This may cause a race condition.

The root cause of the issue remains to be fully investigated. However, it is
clear that not calling RAMCloudTransction.close() is incorrect, as the code is
not thread safe... And actually, as such, it would make sense to actually remove
the finalize() methods from the Java libraries. This *should* be in a future
commit.