Added ImageFileHashPartitioner option to take filename.
Previously ImageFileHashPartitioner only took input from stdin. This was because
image files tend to be compressed, and it was easier to run the image files
through some decompressor and pipe that to the ImageFileHashPartitioner program
than to use some libraries in the program to decompress the files. However there
are use cases where we want the program read an uncompressed image file, and for
this case I added an --inputFile option. If inputFile is not specified, or if it
is set to "-" then input will be read from stdin.