Added TableCreator.
TableCreator creates a specified set of tables in RAMCloud in the order that is
given, with the serverSpan that is given.

The point of TableCreator is to create tables that have specific tableIDs, which
is important for uploading tablets into RAMCloud and making sure that tablet
images actually correspond to a specific server.